requirements.txt
# ======================================
# Tubaroes do Espaco - Requisitos
# NASA Space Apps Challenge 2025
# ======================================

# Core science and arrays
numpy
pandas
scipy

# NetCDF / Xarray
xarray[io]
netcdf4
h5netcdf

# Machine learning
scikit-learn
xgboost
joblib

# Visualization
matplotlib
seaborn
plotly
cartopy

# NASA / Earthdata tooling
earthaccess
requests

# Configuration
pyyaml

# Geospatial
geopandas
rasterio
rioxarray
cartopy
folium

# Performance
dask

# Imaging utilities
pillow
